refact: 검색 페이지 SSR 적용 및 타입 분리
