+++ src/types/search.ts
@@ -1,0 +1,7 @@
+export type SearchResults = {
+  project: any[]
+  study: any[]
+  blog: any[]
+  resume: any[]
+  session: any[]
+}
